+++ R/cNORM.R
@@ -68,42 +68,32 @@
 #' @seealso cNORM.GUI
 #' @examples
 #' # Model internal 'elfe' dataset with the default k = 4 regression on T scores
-#' data.elfe <- prepareData(elfe)
-#' model.elfe <- bestModel(data.elfe)
-#' plotPercentiles(data.elfe, model.elfe)
+#' result <- cnorm(raw = elfe$raw, group = elfe$group)
+#' plotPercentiles(results)
 #'
 #' # Show model fit of models with progressing number of predictors
-#' printSubset(model.elfe)
-#' plotSubset(model.elfe)
+#' printSubset(results)
+#' plotSubset(results)
 #'
 #' # Plot manifest and predicted values, plot series of percentile charts
-#' plotRaw(data.elfe, model.elfe)
+#' plotRaw(results)
 #' \dontrun{
-#' plotPercentileSeries(data.elfe, model.elfe)
+#' plotPercentileSeries(results)
 #' }
 #'
 #' # Additional tests: Check model assumptions
-#' checkConsistency(model.elfe)
-#' plotDerivative(model.elfe)
+#' checkConsistency(results)
+#' plotDerivative(results)
 #'
 #' # Generate norm tables; predict values, here: grade 3.75 from T score 25
 #' # to 75 and within the raw value range of this specific test (0 to 28)
-#' normTable <- normTable(3.75, model.elfe, minNorm=25, maxNorm=75, step=0.5)
-#' rawTable <- rawTable(3.75, model.elfe, minRaw = 0, maxRaw = 28, minNorm=25,
+#' normTable <- normTable(3.75, results, minNorm=25, maxNorm=75, step=0.5)
+#' rawTable <- rawTable(3.75, results, minRaw = 0, maxRaw = 28, minNorm=25,
 #'                      maxNorm=75)
 #'
 #' # Predict a specific norm score
 #' score <- predictNorm(raw = 21, A = 3.75,
-#'                           model = model.elfe, minNorm=25, maxNorm=75)
-#'
-#' # Semi-parametric modeling with Box Cox power transformation for grade 3.75
-#' bcParameters <- boxcox(model.elfe, 3.75)
-#' # Print L, M and S
-#' bcParameters$lambdaBC
-#' bcParameters$meanBC
-#' bcParameters$sdBC
-#' # Plot density function of box cox versus regression model
-#' plotBoxCox(model.elfe, bcParameters, type=2)
+#'                           model = results, minNorm=25, maxNorm=75)
 NULL
 
 
@@ -133,4 +123,69 @@ cNORM.GUI <- function(launch.browser=TRUE){
 
   shiny::runApp(system.file('shiny', package='cNORM'),
                 launch.browser=TRUE)
+}
+
+#' Conducts continuous norming in one step and returns an object including ranked raw data and the continuous
+#' norming model. Please consult the function description ' of 'rankByGroup', 'rankBySlidingWindow' and
+#' 'bestModel' for specifics of the steps in the data preparation and modelling process. Either provide a
+#' grouping vector or an age vector (+ width of the moving window).
+#' @param raw Numeric vector of raw scores
+#' @param group Numeric vector of grouping variable, e. g. grade
+#' @param age Numeric vector with chronological age, please specify width of window
+#' @param width Size of the moving window in case an age vector is used
+#' @param scale type of norm scale, either T (default), IQ, z or percentile (= no
+#' transformation); a double vector with the mean and standard deviation can as well,
+#' be provided f. e. c(10, 3) for Wechsler scale index points
+#' @param descend ranking order (default descent = FALSE): inverses the
+#' ranking order with higher raw scores getting lower norm scores; relevant
+#' for example when norming error scores, where lower scores mean higher
+#' performance
+#' @param weights Vector or variable name in the dataset with weights to compensate imbalances due to insufficient norm
+#' data stratification. All weights have to be numerical and positive. The code to compute weighted percentiles originates from the
+#' Hmisc package (functions) wtd.rank and wtd.table) and is provided by the courtesy of Frank Harrell. Please note, that this
+#' feature is currently EXPERIMENTAL!
+#' @param terms Selection criterion for model building. The best fitting model with
+#' this number of terms is used
+#' @param R2 Adjusted R square as a stopping criterion for the model building
+#' (default R2 = 0.99)
+#' @param k The power constant. Higher values result in more detailed approximations
+#' but have the danger of over-fit (default = 4, max = 6)
+#'
+#' @return object including the ranked raw data and the regression model
+#' @seealso rankByGroup, rankBySlidingWindow, computePowers, bestModel
+#' @examples
+#'
+#' # Using this function with the example dataset 'elfe'
+#' result <- cnorm(raw = elfe$raw, group = elfe$group)
+#'
+cnorm <- function(raw = NULL,
+                  group = NULL,
+                  age = NULL,
+                  scale = "T",
+                  descend = FALSE,
+                  weights = NULL,
+                  width = NA,
+                  k = 4,
+                  terms = 0,
+                  R2 = NULL){
+
+  if(is.numeric(raw)&&is.numeric(group)){
+    if(length(raw)!=length(group)){
+      stop("Please provide numeric vectors of equal length for raw score and group data.")
+    }
+    data <- rankByGroup(raw=raw, group=group, scale=scale, weights=weights, descend = descend)
+  }else if(is.numeric(raw)&&is.numeric(age)&&!is.na(width)){
+    if(length(raw)!=length(age)){
+      stop("Please provide numeric vectors of equal length for raw score and group data.")
+    }
+    data <- rankBySlidingWindow(raw=raw, age=age, scale=scale, weights=weights, descend = descend)
+  }else{
+    stop("Please provide a numerical vector for the raw scores and either a vector for grouping or age of the same length, If you use age, please specify the width of the window.")
+  }
+
+  data <- computePowers(data, k = k)
+  model <- bestModel(data, R2=R2, terms=terms)
+  result <- list(data = data, model = model)
+  class(result) <- "cnorm"
+  return(result)
 }